Refactoring anonymous classes in tests
These anonymous classes in tests contain common code.
I decided to move this common code to a separate class
and take advantage of inheritance for anonymous classes.

But we don't want to define a class in the body of a method.
Therefore, as the base class from which we inherit,
we use also anonymous class.