Fix broken MX-with-fallback test
This test currently fails because subdomain.rubyonrails.org is not
returning an A record. Whether it ever did or not, this is a brittle
test because it's making a real DNS request.

To fix this, and ensure the other tests in this group are also less
brittle, we can stub the lookup responses appropriately for each
example.