Show ruby warnings in CI
I previously removed ruby warnings from the default RSpec options,
because they weren't actionable and were clouding the output.

This reinstates them on CI only. The only warnings currently are from
the mail gem which we don't control, but this will at least allow us to
check that no new warnings are being introduced.

I decided not to reinstate the example profiling that I also removed
earlier, because the build is very quick, so I feel that profiling
output isn't useful.