Fix error assertions for ActiveModel < 6.0
In ActiveModel 5.x, subject.errors[:email] sometimes returns an array of
message strings, sometimes an array of hashes with a message key.

ActiveModel 5.x is installed for ruby 2.4 and below, causing the build
to fail for these rubies. This patch fixes up the various validation
message assertions in the test suite by unwrapping any hashes in the
validation errors before making the assertion.

This workaround can be removed once support is dropped for ActiveModel
5.x.